requirements.txt
streamlit>=1.32.0
langchain>=0.1.13
langchain-community>=0.0.29
langchain-huggingface>=0.0.1
chromadb>=0.6.0
pypdf>=4.1.0
requests>=2.31.0
python-dotenv>=1.0.1
huggingface_hub>=0.21.0
